+++ src/main/java/blackjack/view/Command.java
@@ -1,4 +1,4 @@
-package blackjack.game;
+package blackjack.view;
 
 import java.util.Arrays;
 import java.util.Objects;
@@ -20,7 +20,7 @@ public enum Command {
                 .orElseThrow(() -> new IllegalArgumentException("[ERROR] 존재하지 않는 명령어입니다."));
     }
 
-    public boolean isNo() {
-        return Objects.equals(value, Command.NO.value);
+    public boolean isYes() {
+        return Objects.equals(value, Command.YES.value);
     }
 }
