+++ src/test/java/blackjack/view/CommandTest.java
@@ -1,4 +1,4 @@
-package blackjack.game;
+package blackjack.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
